Add light computation only for first hit, Fix normal in reflected vector

+++ srcs/parse.c
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/23 21:33:22 by craimond          #+#    #+#             */
-/*   Updated: 2024/04/10 16:29:39 by craimond         ###   ########.fr       */
+/*   Updated: 2024/04/10 23:09:01 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,9 +95,9 @@ static void	parse_amblight(t_scene *scene)
 	amblight.brightness = fclamp(ft_atof(ft_strtok(NULL, spaces)), 0, 1);
 	amblight.color = parse_color(ft_strtok(NULL, spaces));
 	amblight.ambient = (t_color){
-        .r = (uint8_t)(amblight.color.r * amblight.brightness),
-        .g = (uint8_t)(amblight.color.g * amblight.brightness),
-        .b = (uint8_t)(amblight.color.b * amblight.brightness),
+        .r = (amblight.color.r * amblight.brightness),
+        .g = (amblight.color.g * amblight.brightness),
+        .b = (amblight.color.b * amblight.brightness),
         .a = 0
     };
 	scene->amblight = amblight;
